Meta: Introduce RecordAssessmentData and deprecate/remove TableData
